realestateseed json

+++ webserver/Utilities/Seeding/CitySeed.cs
@@ -8,6 +8,6 @@ public class CitySeed {
     public class Address{
         public string neighborhood = string.Empty;
         public string[] street = new string[] { string.Empty };
-        public decimal costMultiplyer = (decimal)3.0; //The greater it is, the more expensive real-estates tend to be in that area
+        public decimal costMultiplyer = 3.0m; //The greater it is, the more expensive real-estates tend to be in that area
     }
 }
